🎨 style: follow code convention and eslint
