+++ packages/checkbox/src/components/CheckboxGroup.tsx
@@ -1,13 +1,18 @@
 /** @jsxImportSource @emotion/react */
 import { createClassVariant } from '@jdesignlab/theme';
+import { useMemo } from 'react';
 import { CheckboxGroupContext } from '../context';
 import type { CheckboxGroupProps } from '../types';
 
 export const CheckboxGroup = (props: CheckboxGroupProps) => {
   const { children, defaultValue, ...otherProps } = props;
-  const contextValue = {
-    defaultValues: props.defaultValue ?? []
-  };
+
+  const contextValue = useMemo(
+    () => ({
+      defaultValues: defaultValue ?? []
+    }),
+    [defaultValue]
+  );
 
   return (
     <CheckboxGroupContext.Provider value={contextValue}>
