+++ packages/checkbox/src/context.ts
@@ -1,5 +1,6 @@
 import { createContext } from 'react';
 import type { CheckboxValue } from './types';
+
 export const CheckboxGroupContext = createContext({
   defaultValues: undefined as CheckboxValue
 });
