+++ packages/checkbox/src/types/base.ts
@@ -1,0 +1,2 @@
+export type CheckboxSize = 'sm' | 'md' | 'lg' | 'xl';
+export type CheckboxValue = number[] | string[] | undefined;
