+++ packages/checkbox/src/types/props.ts
@@ -1,14 +1,11 @@
 import type { ColorToken } from '@jdesignlab/theme';
 
-export type CheckboxSize = 'sm' | 'md' | 'lg' | 'xl';
-export type CheckboxValue = number[] | string[] | undefined;
-
 export interface CheckboxProps extends React.InputHTMLAttributes<HTMLInputElement> {
   children?: React.ReactNode;
   color?: ColorToken;
   value?: string | number;
+  // onChange?: () => {};
 }
-
 export interface CheckboxGroupProps {
   children?: React.ReactNode;
   defaultValue?: string[] | number[];
